+++ Changelog.md
@@ -1,11 +1,37 @@
 ## Ignition Fuel Tools 1.x.x
 
+1. Add some AsString functions
+    * [Pull request 54](https://bitbucket.org/ignitionrobotics/ign-fuel-tools/pull-requests/54)
+
+1. Change cache directory structure and use it
+    * [Pull request 57](https://bitbucket.org/ignitionrobotics/ign-fuel-tools/pull-requests/57)
+
 1. Remove server API version from model unique name
     * [Pull request 47](https://bitbucket.org/ignitionrobotics/ign-fuel-tools/pull-requests/47)
+
+1. Client handle unique names and get server info from config
+    * [Pull request 55](https://bitbucket.org/ignitionrobotics/ign-fuel-tools/pull-requests/55)
+
+1. Changed model list to use headers for paging instead of iterating until
+   a 404 is hit. Also added a mechanism to set the user agent
+    * [Pull request 46](https://bitbucket.org/ignitionrobotics/ign-fuel-tools/pull-requests/46)
+
+1. FuelClient: Don't use ServerConfig if there's ModelId
+    * [Pull request 56](https://bitbucket.org/ignitionrobotics/ign-fuel-tools/pull-requests/56)
 
 1. List models from command line
     * [Pull request 43](https://bitbucket.org/ignitionrobotics/ign-fuel-tools/pull-requests/43)
     * [Pull request 48](https://bitbucket.org/ignitionrobotics/ign-fuel-tools/pull-requests/48)
+    * [Pull request 44](https://bitbucket.org/ignitionrobotics/ign-fuel-tools/pull-requests/44)
+
+1. Added const constructors for ModelIter and Model
+    * [Pull request 42](https://bitbucket.org/ignitionrobotics/ign-fuel-tools/pull-requests/42)
+
+1. Use common::URI for server URL
+    * [Pull request 59](https://bitbucket.org/ignitionrobotics/ign-fuel-tools/pull-requests/59)
+
+1. Remove server local name
+    * [Pull request 58](https://bitbucket.org/ignitionrobotics/ign-fuel-tools/pull-requests/58)
 
 ## Ignition Fuel Tools 0.x.x